+++ matchmaker-api/src/main/java/com/example/matchmaker/matchmakerapi/entity/embeddable/GeneroJogoPerfilId.java
@@ -15,5 +15,5 @@ import java.io.Serializable;
 @Embeddable
 public class GeneroJogoPerfilId implements Serializable {
     private Long idPerfil;
-    private Long idGeneroJogos;
+    private Integer idGeneroJogos;
 }
