+++ matchmaker-api/src/main/java/com/example/matchmaker/matchmakerapi/entity/repository/GeneroJogoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GeneroJogoRepository extends JpaRepository<GeneroJogo, Long> {
+public interface GeneroJogoRepository extends JpaRepository<GeneroJogo, Integer> {
 
 }
