+++ matchmaker-api/src/main/java/com/example/matchmaker/matchmakerapi/service/dto/response/GeneroJogoPerfilResponse.java
@@ -11,6 +11,6 @@ import lombok.Setter;
 @Setter
 public class GeneroJogoPerfilResponse {
     private Long perfilId;
-    private Long generoJogoId;
+    private Integer generoJogoId;
     private boolean isVisible;
 }
